fix: script dependance on party order (#6398)
Description
---
fixes script dependence on party order

Motivation and Context
---
Currently, the m-of-n ceremony is dependant on the leader selection to
succeed or not. This fixes the process so that the leader correctly
constructs the input stack for the script in the correct order.

How Has This Been Tested?
---
manual

+++ base_layer/wallet_ffi/src/lib.rs
@@ -9522,19 +9522,19 @@ mod test {
     #[test]
     fn test_emoji_convert() {
         unsafe {
-            let byte = 0 as u8;
+            let byte = 0u8;
             let emoji_ptr = byte_to_emoji(byte);
             let emoji = CStr::from_ptr(emoji_ptr);
 
             assert_eq!(emoji.to_str().unwrap(), EMOJI[0].to_string());
 
-            let byte = 50 as u8;
+            let byte = 50u8;
             let emoji_ptr = byte_to_emoji(byte);
             let emoji = CStr::from_ptr(emoji_ptr);
 
             assert_eq!(emoji.to_str().unwrap(), EMOJI[50].to_string());
 
-            let byte = 125 as u8;
+            let byte = 125u8;
             let emoji_ptr = byte_to_emoji(byte);
             let emoji = CStr::from_ptr(emoji_ptr);
